% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkBookFunctions.R
\name{addCombinedTable}
\alias{addCombinedTable}
\title{Add complete combined data table}
\usage{
addCombinedTable(wb, combTab, sheetNum)
}
\description{
This function adds a worksheet with the combined table showing protein
details for all samples
}
