% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkBookFunctions.R
\name{addFilteredTable}
\alias{addFilteredTable}
\title{Add complete filtered data table}
\usage{
addFilteredTable(wb, filtTab, sheetNum)
}
\description{
This function adds a worksheet with the table showing protein details for all
samples but with the non-specific binding proteins filtered out
}
