% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkBookFunctions.R
\name{addFullTables}
\alias{addFullTables}
\title{Add full data tables for each sample}
\usage{
addFullTables(wb, protData)
}
\description{
This function adds one worksheet for each sample, with the complete protein
details
}
