% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBaitProteinDetection.R
\name{checkBaitDetection}
\alias{checkBaitDetection}
\title{Filter out rows for samples where the bait protein was not detected}
\usage{
checkBaitDetection(sampleTab)
}
\description{
This function loads all the protein data for each sample and then eliminates
rows where the bait protein was not detected.
}
