% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBaitProteinDetection.R
\name{checkProtein}
\alias{checkProtein}
\title{Check that the bait protein is present in the samples protein table}
\usage{
checkProtein(
  SampleName,
  BaitProteinName,
  BaitProteinUniProtID,
  ProteinDat,
  PeptideDat,
  ...
)
}
\description{
This function checks that the bait protein Uniprot ID is present in the 
protein and peptide tables `Accession` column, or that the sample is a 
negative control.
}
