% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePlots.R
\name{coveragePlots}
\alias{coveragePlots}
\title{Make coverage plots for a single sample}
\usage{
coveragePlots(
  SampleName,
  BaitProteinName,
  BaitProteinUniProtID,
  PeptideData,
  ProteinSeq,
  ...
)
}
\description{
This function takes the bait protein sequences and peptide sequence data and
generates a table detailing which amino acids are overlaped by peptides
detected in the sample. It returns a list containing the summary coverage
map, the detailed coverage plot and the sequence for the bait protein.
}
