% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePlots.R
\name{createFastaTable}
\alias{createFastaTable}
\alias{getPosition}
\title{Create fasta table}
\usage{
createFastaTable(sampleTab)

getPosition(peptideSeq, ProteinSeq)
}
\description{
Creates a table with the fasta sequence for each bait protein for the report

Creates a table showing ranges of amino acids in the protein sequence that
are overlapped by the peptides detected in the sample
}
