% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleTable.R
\name{filenametotable}
\alias{filenametotable}
\title{Extract information from file name}
\usage{
filenametotable(fileName)
}
\description{
This function splits the file name and extracts the sample name and protein
}
