% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndFilterDataTables.R
\name{filterCombinedTable}
\alias{filterCombinedTable}
\title{Create filtered protein table}
\usage{
filterCombinedTable(sampleTab)
}
\description{
Read all the protein data tables for the baited samples, combine them into a
single table - add the sample name to the beginning of each header - and add
annotation. Then filter out all proteins that are detected in the negative
controls.
}
