% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProteinData.R
\name{filterMasterProteins}
\alias{filterMasterProteins}
\title{Filter Protein data table for MasterProteins}
\usage{
filterMasterProteins(x)
}
\description{
This function filters on the  columns "Master" in PD v2.4 files to keep only
those rows that are marked "IsMasterProtein". It then removes the "Master"
column as it is no longer needed.
}
