% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeptideTables.R
\name{getPeptideTables}
\alias{getPeptideTables}
\title{Get peptide tables}
\usage{
getPeptideTables(sampleTab)
}
\description{
This function reads each Peptide data file and creates list of peptide level 
data tables -  one for each bait and the control.
}
