% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFunctions.R
\name{loadProteinData}
\alias{loadProteinData}
\title{Load protein data table}
\usage{
loadProteinData(sampleTab, dataDir)
}
\description{
Load protein data. Each table is read into an element of a list, which is 
named according to the sample names.
}
