% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePlots.R
\name{makeCoveragePlots}
\alias{makeCoveragePlots}
\alias{covPlot1}
\alias{covPlot2}
\title{Make summary coverage plot}
\usage{
covPlot1(Protein_seq, features, plotTitle)

covPlot2(Protein_seq, features)

makeCoveragePlots(sampleTab, peptideDat)
}
\description{
This function returns takes the proteins amino acid sequence and the list of
overlapped amino acids (features) and generates a summary plot showing 
the overall coverage across the length of the protein.

This function returns takes the proteins amino acid sequence and the list of
overlapped amino acids (features) and generates a detailed plot showing 
exactly which AAs are overlapped by peptides detected in the sample.

This function retrieves the amino acid sequence for each bait protein, and 
then uses the information in the peptide tables to create two coverage plots
for each bait protein. The first shows the summary of the coverage as a bar,
the second shows the details of which amino acid sequences are covered by 
peptides detected.
}
