% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkBookFunctions.R
\name{makeWorkBook}
\alias{makeWorkBook}
\title{Create the workbook}
\usage{
makeWorkBook(outputFile, plotsList, protData, combData, filtData)
}
\description{
This function creates a new work book and then adds the plots, individual
data tables for each sample, the complete combined data and the filtered
combinded data
}
