% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFunctions.R
\name{selectColumns}
\alias{selectColumns}
\title{Select columns from Protein data table}
\usage{
selectColumns(x)
}
\description{
This function selects the columns to keep from the PD protein table.
}
