% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndFilterDataTables.R
\name{combineProteinTables}
\alias{combineProteinTables}
\title{Combine all protein data into a single table}
\usage{
combineProteinTables(sampleTab, annotTab)
}
\description{
Read all the protein data tables, combine them into a single table - add the
sample name to the beginning of each header - and add annotation.
}
