% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAminoAcidSequences.R
\name{createFastaTable}
\alias{createFastaTable}
\title{Create fasta table}
\usage{
createFastaTable(sampleTab)
}
\description{
Creates a table with the fasta sequence for each bait protein for the report
}
