% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create the Simple RIME report}
\usage{
createReport(
  sampleTable,
  dataDir,
  additionalFasta = NULL,
  outputFileName = NULL
)
}
\arguments{
\item{sampleTable}{data.frame; the sample table}

\item{dataDir}{character; Path to directory containing the Proteome
Discoverer output files}

\item{additionalFasta}{character; a vector of paths to custom fasta files to
be included in the annotation (optional)}

\item{outputFileName}{charcter; an output file name (optional)}
}
\description{
This function generates report as an Excel workbook. The input is a sample
table.
}
\examples{

}
