% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create the Simple RIME report}
\usage{
createReport(
  sampleTable,
  dataDir,
  additionalFasta = NULL,
  outputFileName = NULL
)
}
\arguments{
\item{sampleTable}{data.frame; the sample table}

\item{dataDir}{character; Path to directory containing the Proteome
Discoverer output files}

\item{additionalFasta}{character; a vector of paths to custom fasta files to
be included in the annotation (optional)}

\item{outputFileName}{charcter; an output file name (optional)}
}
\description{
This function generates report as an Excel workbook. The input is a sample
table.
}
\details{
The sample table should contain 5 columns:  
\itemize{
 \item{ ProteinFile - the file name for 'protiengroups' data }
 \item{ PeptideFile - the file name for 'psms' data }
 \item{ SampleName - the name for the sample - this is used for the plot
                     titles and the names of the worksheets }
 \item{ BaitProteinName - the protein name for the bait protein, e.g. ER,
                          FOXA1 etc. This is only used in the plot titles,
                          it could be anything }
 \item{ BaitProteinUniProtID - the UniProt ID for the bait protein, e.g.
                               P03372, P55317 this is used to look up the
                               fasta sequence }
}
The proteins are filtered on the "Master" column, if present, to only keep
proteins marked "IsMasterProtein".
}
\examples{

}
