% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleTable.R
\name{createSampleTable}
\alias{createSampleTable}
\title{Create the sample table based on files in the data directory}
\usage{
createSampleTable(dataDir)
}
\arguments{
\item{dataDir}{character; Path to directory containing the Proteome
Discoverer output files}
}
\value{
A sample table suitable for input into the `createReport` function. A
  copy of the table it also written to the data directory as a comma 
  separated table.
}
\description{
This function scans the contents of a directory for the presence of Proteome
Discoverer results files.
}
\details{
The files required for each sample are a Protein table and a Peptide table.
The filenames are expected to have the following naming convention:

QE_HF_<Date>_<PR_number>_<Initials>_<SampleName>_<BaitProteinName>_Proteins.txt
QE_HF_<Date>_<PR_number>_<Initials>_<SampleName>_<BaitProteinName>_PeptideGroups.txt

The function creates a new file, in the sample directory, that contains a
sample table, which is used to generate the report. 

The sample table will contain 5 columns:
\itemize{
 \item{ ProteinFile - the file name for 'protiengroups' data }
 \item{ PeptideFile - the file name for 'psms' data }
 \item{ SampleName - the name for the sample - this is used for the plot
                     titles and the names of the worksheets }
 \item{ BaitProteinName - the protein name for the bait protein, e.g. ER,
                          FOXA1 etc. This is only used in the plot titles,
                          it could be anything }
 \item{ BaitProteinUniProtID - the UniProt ID for the bait protein, e.g.
                               P03372, P55317 this is used to look up the
                               fasta sequence }
}

The script will attempt to deduce the SampleName and BaitProteinName from the
the file name. It will also try to determine the UniProt ID from the bait
protein name. If the BaitProteinName is associated with multiple UniProt IDs,
the script will return all possibilities.

It is important to check the details contained in the table, and modify it if
necessary, prior to generating the report. In particular check the bait
protein name and UniProt ID. The negative control should be called "Control",
"Empty" or "IgG" and have no UniProtID.
}
\examples{

}
