% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAndFilterDataTables.R
\name{getNonSpecificProteins}
\alias{getNonSpecificProteins}
\alias{addSampleName}
\title{Add sample name to column header}
\usage{
addSampleName(tab, sampleName)

getNonSpecificProteins(sampleTab)
}
\description{
Add the sample name to all column headers except "Accession"

Read all protein tables for negative controls and compile a vector of 
the proteins detected. These will be considered as present due to 
non-specific binding and removed from the final filtered protein table.
}
