% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinTables.R
\name{getProteinTables}
\alias{getProteinTables}
\title{Get protein tables}
\usage{
getProteinTables(sampleTab)
}
\description{
This function reads each Protein data file and creates list of protein level 
data tables with annotations -  one for each bait and the control.
}
