% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCustomFasta.R
\name{loadCustomFasta}
\alias{loadCustomFasta}
\title{Add data from custom fasta to annotation}
\usage{
loadCustomFasta(additionalFasta)
}
\description{
This function reads annotations and sequences from fasta files and adds them
to the annotation used for the report.
The Protein accession, Gene ID and description are taken from the name for 
each sequence. the expected format is
   >sp|<Accession>|<GeneID> <Description>
e.g.
   >sp|NUP98-HOXA9|NUP98-HOXA9_HUMAN OS=Homo sapiens
where the Accession is NUP98-HOXA9 and the Gene ID is NUP98-HOXA9_HUMAN. Any
further information is add to the Description.
}
